iio: adc: nexell_adc: stabilizes sampling value
This patch provides a bug work-around code for nxp4330 ADC.

When the requested adc sampling is finished, there is a problem that
the next sampling is automatically started and the wrong result is read.

So I have a proprietary cpu and I applied address filtering,
but the bus transaction still occasionally bounces.

This patch further minimizes fetch time by lowering the sampling rate,
and attempts to finish the transaction quickly using the dsb command
when polling.

Change-Id: I895cfe4e2ec3d711092dcd178cbe3722cb057798
Signed-off-by: Bon-gyu, KOO <freestyle@nexell.co.kr>