misc: backward-camera:remove the tearing symptom of the progress sensor
This patch adds a source to eliminate the tearing symptoms of the
progress sensor. This symptom is generated by processing the input image
through one buffer, and it is changed so that it can be processed by
multiple buffers.

To enable the processing of multiple buffers,
set as follows.

	in the device.c file
	- file path : kernel/arch/arm/plat-s5p6818/avn_ref/device.c

static struct nxp_backward_camera_platform_data
backward_camera_plat_data = {
	.....
	.multi_buffer           = true,
	.....
}

Change-Id: I3e80de807e584b62642708f41123a584cc2a86c6
Signed-off-by: JongKeun Choi <jkchoi@nexell.co.kr>